Add Next, Until, Previously, and Since to grammar
These are some additional operators useful for the l2s construction.
Next and Until are the primitive "future" temporal operators (which we
can translate the current prime/always/eventually into), and
Previously and Since are their past analogues that will only be used for
the construction.

The Next and Previously modalities are written `X` and `X^-1`, as prefix
operators. Until and Since are infix `until` and `since`.

Signed-off-by: Tej Chajed <tchajed@vmware.com>